+++ Sources/SkyFloatingLabelTextField.swift
@@ -39,9 +39,9 @@ open class SkyFloatingLabelTextField: UITextField { // swiftlint:disable:this ty
     // MARK: Animation timing
 
     /// The value of the title appearing duration
-    dynamic open var titleFadeInDuration: TimeInterval = 0.2
+    @objc dynamic open var titleFadeInDuration: TimeInterval = 0.2
     /// The value of the title disappearing duration
-    dynamic open var titleFadeOutDuration: TimeInterval = 0.3
+    @objc dynamic open var titleFadeOutDuration: TimeInterval = 0.3
 
     // MARK: Colors
 
@@ -67,23 +67,34 @@ open class SkyFloatingLabelTextField: UITextField { // swiftlint:disable:this ty
     }
 
     /// A UIFont value that determines text color of the placeholder label
-    dynamic open var placeholderFont: UIFont? {
+    @objc dynamic open var placeholderFont: UIFont? {
         didSet {
             updatePlaceholder()
         }
     }
 
     fileprivate func updatePlaceholder() {
-        if let placeholder = placeholder, let font = placeholderFont ?? font {
-                attributedPlaceholder = NSAttributedString(
-                    string: placeholder,
-                    attributes: [NSForegroundColorAttributeName: placeholderColor, NSFontAttributeName: font]
-                )
+        guard let placeholder = placeholder, let font = placeholderFont ?? font else {
+            return
         }
+        let color = isEnabled ? placeholderColor : disabledColor
+        #if swift(>=4.0)
+            attributedPlaceholder = NSAttributedString(
+                string: placeholder,
+                attributes: [
+                    NSAttributedStringKey.foregroundColor: color, NSAttributedStringKey.font: font
+                ]
+            )
+        #else
+            attributedPlaceholder = NSAttributedString(
+                string: placeholder,
+                attributes: [NSForegroundColorAttributeName: color, NSFontAttributeName: font]
+            )
+        #endif
     }
 
     /// A UIFont value that determines the text font of the title label
-    dynamic open var titleFont: UIFont = .systemFont(ofSize: 13) {
+    @objc dynamic open var titleFont: UIFont = .systemFont(ofSize: 13) {
         didSet {
             updateTitleLabel()
         }
@@ -107,6 +118,14 @@ open class SkyFloatingLabelTextField: UITextField { // swiftlint:disable:this ty
     @IBInspectable dynamic open var errorColor: UIColor = .red {
         didSet {
             updateColors()
+        }
+    }
+    
+    /// A UIColor value that determines the color used for the title label and line when text field is disabled
+    @IBInspectable dynamic open var disabledColor: UIColor = UIColor(white: 0.88, alpha: 1.0) {
+        didSet {
+            updateControl()
+            updatePlaceholder()
         }
     }
 
@@ -289,7 +308,7 @@ open class SkyFloatingLabelTextField: UITextField { // swiftlint:disable:this ty
     /**
      Invoked when the editing state of the textfield changes. Override to respond to this change.
      */
-    open func editingChanged() {
+    @objc open func editingChanged() {
         updateControl(true)
         updateTitleLabel(true)
     }
@@ -350,6 +369,14 @@ open class SkyFloatingLabelTextField: UITextField { // swiftlint:disable:this ty
         return result
     }
 
+    /// update colors when is enabled changed
+    override open var isEnabled: Bool {
+        didSet {
+            updateControl()
+            updatePlaceholder()
+        }
+    }
+    
     // MARK: - View updates
 
     fileprivate func updateControl(_ animated: Bool = false) {
@@ -375,7 +402,9 @@ open class SkyFloatingLabelTextField: UITextField { // swiftlint:disable:this ty
     }
 
     fileprivate func updateLineColor() {
-        if hasErrorMessage {
+        if !isEnabled {
+            lineView.backgroundColor = disabledColor
+        } else if hasErrorMessage {
             lineView.backgroundColor = errorColor
         } else {
             lineView.backgroundColor = editingOrSelected ? selectedLineColor : lineColor
@@ -383,7 +412,9 @@ open class SkyFloatingLabelTextField: UITextField { // swiftlint:disable:this ty
     }
 
     fileprivate func updateTitleColor() {
-        if hasErrorMessage {
+        if !isEnabled {
+            titleLabel.textColor = disabledColor
+        } else if hasErrorMessage {
             titleLabel.textColor = errorColor
         } else {
             if editingOrSelected || isHighlighted {
@@ -395,7 +426,9 @@ open class SkyFloatingLabelTextField: UITextField { // swiftlint:disable:this ty
     }
 
     fileprivate func updateTextColor() {
-        if hasErrorMessage {
+        if !isEnabled {
+            super.textColor = disabledColor
+        } else if hasErrorMessage {
             super.textColor = errorColor
         } else {
             super.textColor = cachedTextColor