+++ Sources/SkyFloatingLabelTextFieldWithIcon.swift
@@ -60,7 +60,7 @@ open class SkyFloatingLabelTextFieldWithIcon: SkyFloatingLabelTextField {
     open var iconLabel: UILabel!
 
     /// A UIFont value that determines the font that the icon is using
-    dynamic open var iconFont: UIFont? {
+    @objc dynamic open var iconFont: UIFont? {
         didSet {
             iconLabel?.font = iconFont
         }
@@ -222,7 +222,9 @@ open class SkyFloatingLabelTextFieldWithIcon: SkyFloatingLabelTextField {
     }
 
     fileprivate func updateIconLabelColor() {
-        if self.hasErrorMessage {
+        if !isEnabled {
+            iconLabel?.textColor = disabledColor
+        } else if hasErrorMessage {
             iconLabel?.textColor = errorColor
         } else {
             iconLabel?.textColor = editingOrSelected ? selectedIconColor : iconColor